requirements.txt
openai
transformers
questionary
matplotlib
python-dotenv
ollama 